Make tag matching case-insensitive
Currently smokey is failing because all the most recent annotations
tagged matching a query for ?tags=hypothesis have the tag "Hypothesis"
with an upper-case h.

For now this will fix the tests by asserting that each annotation has at
least one tag that is a case-insensitive match for the specified tag,
but in the future we may with to revisit how detailed these tests are.

Certainly, I don't want us to have to reimplement all of the
analysis/tokenisation that ElasticSearch does just to get these tests to
pass in the general case.