Reimplement checkrealtime as part of smokey
This reimplements a test that the websocket streamer is working
correctly end-to-end, as currently implemented by the checkrealtime
tool, also in this repository.

Checkrealtime is lovely, but it's written in Go, is a faff to maintain,
and currently doesn't support retrying the API calls it makes.

This code uses the `websockets` library, which makes asynchronous
interaction with a websocket a breeze, and will now take advantage of
the default retry settings of the requests session.