Rewrite websocket notification steps
These changes, which move the connection to the websocket into a
separate process which communicates with the tests through a
multiprocessing.Queue, hopefully fix the concurrency bug that results in
intermittent smokey failures.

As a bonus, the code is slightly more conventional, easier to read, and
there's less crazy wait_for stuff that was only there to work around
using asyncio in a synchronous environment.