Fix race condition in websocket listener
This test was intermittently failing because the "I am listening for
notifications on the websocket" step returned before the websocket
listener (in a child process) was set up and listening.

We fix this by adding another `multiprocessing.Event` to signal
readiness from the remote process when the websocket connection is
established and configured.