Fix a deadlock by draining the WS messages queue
Every once in a while, Smokey would fail to exit cleanly, with no
problems reported to the logs. Eventually, I straced the "stuck"
processing and found the following:

     ubuntu@mon:~$ sudo strace -p 24713
     Process 24713 attached
     write(6, "\0\0\7t\200\3Xj\7\0\0{\"type\": \"annotation-"...,
     1912^CProcess 24713 detached

The latter visible part of the blocked `write(2)` call here is
recognisable as the JSON sent down the websocket to notify of an
annotation event. FD 6 for this process was a pipe:

     ubuntu@mon:~$ sudo ls -dl /proc/24713/fd/6
     l-wx------ 1 libuuid libuuid 64 Feb 18 12:35 /proc/24713/fd/6 -> pipe:[4553700]

Thinking about this for a while, it became apparent that the most likely
scenario was that the blocked write call corresponded to a `put()`
method call on a `multiprocessing.Queue` object. A minimal
reproduction[1] demonstrates that an apparently non-blocking queue write
can block if the underlying pipe is full.

[1]: https://gist.github.com/b6f9a6c5196a06cc3c67a367458506ab

So, the fix is (hopefully) straightforward -- ensure that the queue is
drained before attempting to join the process using it.