Pre-emptively fix a possible race condition
The previous commit fixes the most obvious problem with blocked queue
writes, but potentially left a race condition where the queue can have
data put into it between draining the queue and calling
`websocket.join()`.

This commit reorders operations in the inner loop of the queue writer
process so that by the time queue draining starts, we can guarantee that
no more messages will be written to the queue.

In addition, we now explicitly close the queue in the writer process so
that any further attempts to write to it will result in an exception.