Run Smokey tests with a timeout
The tests involve making many HTTP requests to external services doing
complex things like automating web browsers, so there are potentially
many ways that a test could get stuck.

To ensure the tests complete in a reasonable amount of time no matter
what, add a 4 minute timeout. We currently run the tests every 5 minutes
on the monitoring infrastructure, so this ensures that the previous run
will have completed when the next one is started.

For as-yet undetermined reasons, the timeout only works if the _shell_
instead of _exec_ form of CMD is used.